Bundles

We try to keep the same syntax for AbsBundle and TnBundle. Note that not all methods are available for TnBundle.

Construct bundles

IntersectionTheory.bundle — Function

bundle(X::AbsVariety, ch)
bundle(X::AbsVariety, r, c)

Construct a bundle on $X$ by specifying its Chern character, or its rank and total Chern class.

IntersectionTheory.OO — Function

OO(X::AbsVariety)
OO(X::TnVariety)

Return the trivial bundle $\mathcal O_X$ on $X$.

OO(X::AbsVariety, n)
OO(X::AbsVariety, D)

Return the line bundle $\mathcal O_X(n)$ on $X$ if $X$ has been given a polarization, or a line bundle $\mathcal O_X(D)$ with first Chern class $D$.

IntersectionTheory.bundles — Function

bundles(X::AbsVariety)
bundles(X::TnVariety)

Return the tautological bundles of a variety $X$. Same as X.bundles.

IntersectionTheory.tangent_bundle — Method

tangent_bundle(X::AbsVariety)
tangent_bundle(X::TnVariety)

Return the tangent bundle of a variety $X$. Same as X.T.

IntersectionTheory.cotangent_bundle — Method

cotangent_bundle(X::AbsVariety)
cotangent_bundle(X::TnVariety)

Return the cotangent bundle of a variety $X$.

IntersectionTheory.canonical_bundle — Function

canonical_bundle(X::AbsVariety)

Return the canonical bundle of a variety $X$.

IntersectionTheory.canonical_class — Function

canonical_class(X::AbsVariety)

Return the canonical class of a variety $X$.

Examples

julia> OO(proj(2))
AbsBundle of rank 1 on AbsVariety of dim 2

julia> bundles(grassmannian(2, 4))
2-element Vector{IntersectionTheory.AbsBundle}:
 AbsBundle of rank 2 on AbsVariety of dim 4
 AbsBundle of rank 2 on AbsVariety of dim 4

julia> bundles(grassmannian(2, 4, bott=true))
2-element Vector{IntersectionTheory.TnBundle}:
 TnBundle of rank 2 on TnVariety of dim 4
 TnBundle of rank 2 on TnVariety of dim 4

julia> tangent_bundle(variety(3))
AbsBundle of rank 3 on AbsVariety of dim 3

Characteristic classes

IntersectionTheory.ch — Function

ch(F::AbsBundle)

Return the Chern character.

IntersectionTheory.chern — Method

chern(F::AbsBundle)
chern(F::TnBundle)

Compute the total Chern class.

IntersectionTheory.chern — Method

chern(k::Int, F::AbsBundle)
chern(k::Int, F::TnBundle)

Compute the $k$-th Chern class.

IntersectionTheory.ctop — Function

ctop(F::AbsBundle)
ctop(F::TnBundle)

Compute the top Chern class.

IntersectionTheory.segre — Function

segre(F::AbsBundle)

Compute the total Segre class.

segre(k::Int, F::AbsBundle)

Compute the $k$-th Segre class.

IntersectionTheory.chi — Function

chi(F::AbsBundle)
chi(F::AbsBundle, G::AbsBundle)

Compute the holomorphic Euler characteristic $\chi(F)$, or the Euler pairing $\chi(F,G)$.

IntersectionTheory.todd — Method

todd(F::AbsBundle)

Compute the Todd class.

IntersectionTheory.pontryagin — Method

pontryagin(F::AbsBundle)

Compute the total Pontryagin class.

IntersectionTheory.pontryagin — Method

pontryagin(k::Int, F::AbsBundle)

Compute the $k$-th Pontryagin class.

Note

For a TnBundle of rank $r$, since there are no Chow ring present, its Chern classes are represented as formal polynomials in $r$ variables. Arithmetics are only available for Chern classes of a same TnBundle. See examples below.

Examples

For AbsBundle.

julia> X, (F,) = variety(2, [2=>"c"]);

julia> ch(F)
1//2*c₁^2 + c₁ - c₂ + 2

julia> chern(F)
c₁ + c₂ + 1

julia> pontryagin(F)
c₁^2 - 2*c₂ + 1

julia> chern(proj(2))
3*h^2 + 3*h + 1

julia> chern(2, proj(2))
3*h^2

julia> todd(proj(2))
h^2 + 3//2*h + 1

julia> chi(cotangent_bundle(proj(2)))
-1

julia> G = grassmannian(2, 4); S, Q = bundles(G);

julia> chern(1, S)
c₁

julia> integral(chern(1, S)^4)
2

julia> chern(1, S) * chern(2, Q)
c₁*c₂

For TnBundle.

julia> G = grassmannian(2, 4, bott=true); S, Q = bundles(G);

julia> chern(1, S)
Chern class c₁ of TnBundle of rank 2 on TnVariety of dim 4

julia> integral(chern(1, S)^4)
2

julia> chern(1, S) * chern(2, Q) # this will not work
ERROR: AssertionError: a.F == b.F

Operators

IntersectionTheory.symmetric_power — Function

symmetric_power(k, F::AbsBundle)
symmetric_power(k::Int, F::TnBundle)

Return the $k$-th symmetric power. For an AbsBundle, $k$ can contain parameters.

IntersectionTheory.exterior_power — Function

exterior_power(k::Int, F::AbsBundle)
exterior_power(k::Int, F::TnBundle)

Return the $k$-th exterior power.

LinearAlgebra.det — Function

det(F::AbsBundle)
det(F::TnBundle)

Return the determinant bundle.

IntersectionTheory.dual — Function

dual(F::AbsBundle)
dual(F::TnBundle)

Return the dual bundle.

IntersectionTheory.schur_functor — Function

schur_functor(λ::Vector{Int}, F::AbsBundle)
schur_functor(λ::Partition, F::AbsBundle)

Return the result of the Schur functor $\mathbf S^\lambda$.