Chern number computations

We gather some Chern numbers related computations here.

Cobordism ring

Let $\Omega$ be the (complex) cobordism ring with rational coefficients. Milnor showed that $\Omega$ is a polynomial ring freely generated by cobordism classes of projective spaces. We implemented the interface CobordRing and CobordRingElem to model $\Omega$ and its elements. The main application will be the computation of Chern numbers for Hilbert schemes of points on a surface and for generalized Kummer varieties.

julia> Ω = IntersectionTheory.Omega
Cobordism Ring

julia> Ω[1]
[P1]

julia> (1 + Ω[1])^2 + Ω[2]
1 + 2*[P1] + [P2] + [P1xP1]

julia> Nemo.PolynomialRing(Ω, ["z"])[1]
Multivariate Polynomial Ring in z over Cobordism Ring

The letter $\Omega$ can be typed using \Omega[tab] in Julia, and of course one can use other letters.

We can plug in an AbsVariety or a TnVariety to construct the cobordism class.

julia> K3 = complete_intersection(proj(3), 4)
AbsVariety of dim 2

julia> Ω(K3)
-16*[P2] + 18*[P1xP1]

julia> Ω(flag(1,2,3))
-3//2*[P3] + 4*[P2xP1] - 3//2*[P1xP1xP1]

Each cobordism class is uniquely determined by its Chern numbers and vice versa, so we can perform conversions between a cobordism class and a dictionary of Chern numbers.

julia> chern_numbers(Ω(K3))
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 2 entries:
  1₂ => 0
  2₁ => 24

julia> Ω(chern_numbers(K3))
-16*[P2] + 18*[P1xP1]

Integral of a top-degree class depends only on the Chern numbers, hence we can perform integration against a cobordism class. We also show the usual integration in Chow ring as a comparison.

julia> integral(Ω[4], todd(4)),                     integral(todd(proj(4)))
(1, 1)

julia> integral(Ω[2]^2, sqrt(todd(variety(4)))),    integral(sqrt(todd(proj(2)*proj(2))))
(49//1024, 49//1024)

julia> integral(Ω(flag(1,2,3)), ch(variety(3).T)),  integral(ch(flag(1,2,3).T))
(-1, -1)

Genus on the cobordism ring

An $R$-valued genus $\phi$ is a ring morphism from $\Omega$ to another ring $R$. The Todd genus is a such example with $R=\mathbf Q$. Milnor showed that genera are in one-to-one correspondence with formal power series in $R[[z]]$ with constant coefficient 1: for an $n$-dimensional complex manifold $X$, write $\gamma_i$ with $1\le i\le n$ for the Chern roots of the tangent bundle of $X$, then the power series $f_\phi\in R[[z]]$ satisfies

\[\phi([X]) = \int_X\prod_{i=1}^nf_\phi(\gamma_i).\]

For example, the power series for the Todd genus is the Taylor expansion of $\frac{z}{1-\exp(-z)}$. One can also expand the product in the integrand in terms of Chern classes of $X$: this gives one expression in each degree $n$. We will refer to the sum of these expressions as the total class of the genus. For Todd genus, we get the familiar

\[1+\tfrac12c_1+\tfrac1{12}(c_1^2+c_2)+\cdots\]

which we can obtain using todd(n).

IntersectionTheory.universal_genus — Function

universal_genus(n::Int)
universal_genus(n::Int, t::Int)

Compute the total class of the universal genus up to degree $n$, possibly with a twist $t$, in terms of the Chern classes.

One particular interesting genus is the identity $\mathrm{id}:\Omega \to \Omega$, which we will call the universal genus. One would get a formal power series $f_{\mathrm{id}}$ in $\Omega[[z]]$. We can compute the first terms of this power series, as well as the total class of the genus. We also verify that by taking the integral we indeed get the identity.

julia> IntersectionTheory._taylor(4)
5-element Vector{IntersectionTheory.CobordRingElem}:
 1
 1//2*[P1]
 1//3*[P2] - 1//4*[P1xP1]
 1//4*[P3] - 1//2*[P2xP1] + 1//4*[P1xP1xP1]
 1//5*[P4] - 1//2*[P3xP1] - 2//9*[P2xP2] + 5//6*[P2xP1xP1] - 5//16*[P1xP1xP1xP1]

julia> universal_genus(2)
1 + 1//2*[P1]*c₁ + (1//3*[P2] - 1//4*[P1xP1])*c₁^2 + (-2//3*[P2] + 3//4*[P1xP1])*c₂

julia> integral(Ω[2], universal_genus(2))
[P2]

One can also twist a genus $\phi$ by using the Chern character of the anti-canonical bundle, and obtain a new genus $\phi_t$.

Here we compute the total class for the twisted genus $\mathrm{id}_1$.

julia> universal_genus(2, 1)
1 + (1 + 1//2*[P1])*c₁ + (1//2 + 1//2*[P1] + 1//3*[P2] - 1//4*[P1xP1])*c₁^2 + (-2//3*[P2] + 3//4*[P1xP1])*c₂

Chern numbers of the Hilbert scheme of a surface

Let $S$ be a smooth projective surface. Ellingsrud–Göttsche–Lehn showed that the cobordism class of the Hilbert scheme $S^{[n]}$ of $n$ points on $S$ depends only on the cobordism class of $S$. In other words, the Chern numbers of $S^{[n]}$ only depends on the Chern numbers of $S$, $c_1(S)^2$ and $c_2(S)$.

On the other hand, the cobordism classes of surfaces are linearly generated by $[\mathbf P^2]$ and $[\mathbf P^1\times \mathbf P^1]$. For these two surfaces, their Hilbert schemes admit natural torus actions, so their Chern numbers can be readily computed using Bott's formula.

We provide the constructors for $(\mathbf P^2)^{[n]}$ and $(\mathbf P^1\times \mathbf P^1)^{[n]}$ as TnVariety, which will allow the computation of their Chern numbers using Bott's formula. Then we implement a function that computes the cobordism class of $S^{[n]}$ for an arbitrary cobordism class $[S]$ specified by its two Chern numbers $c_1^2$ and $c_2$, using the strategy given in loc. cit. Essentially, it makes use of the generating series

\[H(S):=\sum_{n=0}^\infty[S^{[n]}]z^n\in\Omega[[z]].\]

IntersectionTheory.hilb_P2 — Function

hilb_P2(n::Int)

Construct the Hilbert scheme of $n$ points on $\mathbf P^2$ as a TnVariety.

IntersectionTheory.hilb_P1xP1 — Function

hilb_P1xP1(n::Int)

Construct the Hilbert scheme of $n$ points on $\mathbf P^1\times\mathbf P^1$ as a TnVariety.

IntersectionTheory.hilb_surface — Function

hilb_surface(n::Int, c1_2::Int, c2::Int)

Construct the cobordism class of the Hilbert scheme of $n$ points on a surface with given Chern numbers $c_1^2$ and $c_2$.

IntersectionTheory.hilb_K3 — Function

hilb_K3(n::Int)

Construct the cobordism class of a hyperkähler variety of $\mathrm{K3}^{[n]}$-type.

Examples

julia> X = hilb_P2(2)
TnVariety of dim 4

julia> chern_numbers(X)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 5 entries:
  2₂   => 90
  2₁1₂ => 162
  3₁1₁ => 54
  1₄   => 243
  4₁   => 9

julia> hilb_surface(2, 0, 24)
72*[P4] - 225*[P3xP1] + 84*[P2xP2] + 36*[P2xP1xP1] + 36*[P1xP1xP1xP1]

julia> hilb_K3(2)
72*[P4] - 225*[P3xP1] + 84*[P2xP2] + 36*[P2xP1xP1] + 36*[P1xP1xP1xP1]

julia> chern_numbers(hilb_K3(2), nontriv=true)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 2 entries:
  2₂ => 828
  4₁ => 324

We record the Chern numbers of $\mathrm{K3}^{[n]}$ with $n\in\{5,6,7\}$, since these are not found in the literature.

julia> chern_numbers(hilb_K3(5), nontriv=true)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 7 entries:
  10₁  => 176256
  6₁4₁ => 5075424
  6₁2₂ => 12168576
  8₁2₁ => 1774080
  4₁2₃ => 52697088
  2₅   => 126867456
  4₂2₁ => 21921408

julia> chern_numbers(hilb_K3(6), nontriv=true)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 11 entries:
  6₁4₁2₁ => 392090040
  8₁4₁   => 59314272
  8₁2₂   => 139942280
  12₁    => 1073720
  6₁2₃   => 927397840
  10₁2₁  => 14450680
  4₃     => 697106648
  2₆     => 9277276480
  4₁2₄   => 3910848640
  4₂2₂   => 1650311720
  6₂     => 93495320

julia> chern_numbers(hilb_K3(7), nontriv=true)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 15 entries:
  8₁6₁   => 1296158976
  10₁4₁  => 569044224
  10₁2₂  => 1324608768
  6₁2₄   => 79324710912
  6₁4₂   => 14528215296
  8₁4₁2₁ => 5296568832
  6₁4₁2₂ => 33935583744
  4₃2₁   => 59674012416
  2₇     => 765374164992
  4₁2₅   => 326732507136
  6₂2₁   => 8273055744
  4₂2₃   => 139582386432
  12₁2₁  => 102477312
  8₁2₃   => 12357114624
  14₁    => 5930496

Chern numbers of the generalized Kummer varieties

By results of Nieper-Wißkirchen, the cobordism classes of generalized Kummer varieties can be obtained using the generating series $H(S)$ for one surface $S$ with $c_1(S)^2\ne0$, along with the values of $H(S)$ under the twisted genus $\mathrm{id}_1$ and $\mathrm{id}_{-1}$. We take $S$ to be the projective plane $\mathbf P^2$.

IntersectionTheory.generalized_kummer — Function

generalized_kummer(n::Int)

Construct the cobordism class of a hyperkähler variety of $\mathrm{Kum}_{n}$-type.

Examples

julia> generalized_kummer(2)
216*[P4] - 675*[P3xP1] - 132*[P2xP2] + 972*[P2xP1xP1] - 378*[P1xP1xP1xP1]

julia> chern_numbers(generalized_kummer(2), nontriv=true)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 2 entries:
  2₂ => 756
  4₁ => 108

A polynomial defined by Libgober–Wood

Libgober–Wood defined a universal polynomial with coefficients in Chern numbers of an $n$-dimensional compact Kähler manifold, which can be used to obtain restrictions on certain topological invariants. They are computed up to dimension 9 in this survey.

We have a function to compute this polynomial. The dimension-10 one can be computed in less than 1 second, and the dimension-16 one in about 1 minute.

Of course one can also compute the value of this polynomial for a specific variety.

IntersectionTheory.libgober_wood_polynomial — Function

libgober_wood_polynomial(n::Int)
libgober_wood_polynomial(X::AbsVariety)

Compute the polynomial defined by Libgober–Wood in dimension $n$.

Examples

julia> libgober_wood_polynomial(3)
-1//24*c₁*c₂*z^3 + (1//12*c₁*c₂ + 1//2*c₃)*z^2 - 3//2*c₃*z + c₃

julia> libgober_wood_polynomial(proj(6))
z^6 - 7*z^5 + 21*z^4 - 35*z^3 + 35*z^2 - 21*z + 7

The degree-2 term is shown to depend on only two Chern numbers.

julia> [Nemo.coeff(libgober_wood_polynomial(n), 2) for n in 1:10]
10-element Vector{IntersectionTheory.ChRingElem}:
 0
 1//12*c₁^2 + 1//12*c₂
 1//12*c₁*c₂ + 1//2*c₃
 1//12*c₁*c₃ + 7//6*c₄
 1//12*c₁*c₄ + 25//12*c₅
 1//12*c₁*c₅ + 13//4*c₆
 1//12*c₁*c₆ + 14//3*c₇
 1//12*c₁*c₇ + 19//3*c₈
 1//12*c₁*c₈ + 33//4*c₉
 1//12*c₁*c₉ + 125//12*c₁₀

We observe that there should be similar results for the degree-4 term.

julia> [Nemo.coeff(libgober_wood_polynomial(n), 4) for n in 1:10]
10-element Vector{IntersectionTheory.ChRingElem}:
 0
 0
 0
 -1//720*c₁^4 + 1//180*c₁^2*c₂ + 1//240*c₂^2 + 1//720*c₁*c₃ - 1//720*c₄
 -1//720*c₁^3*c₂ + 1//240*c₁*c₂^2 + 1//720*c₁^2*c₃ + 29//720*c₁*c₄ + 1//24*c₅
 -1//720*c₁^3*c₃ + 1//240*c₁*c₂*c₃ + 1//720*c₁^2*c₄ - 1//240*c₃^2 + 1//240*c₂*c₄ + 23//240*c₁*c₅ + 31//120*c₆
 -1//720*c₁^3*c₄ + 1//240*c₁*c₂*c₄ + 1//720*c₁^2*c₅ - 1//240*c₃*c₄ + 1//240*c₂*c₅ + 31//180*c₁*c₆ + 301//360*c₇
 -1//720*c₁^3*c₅ + 1//240*c₁*c₂*c₅ + 1//720*c₁^2*c₆ - 1//240*c₃*c₅ + 1//240*c₂*c₆ + 97//360*c₁*c₇ + 81//40*c₈
 -1//720*c₁^3*c₆ + 1//240*c₁*c₂*c₆ + 1//720*c₁^2*c₇ - 1//240*c₃*c₆ + 1//240*c₂*c₇ + 31//80*c₁*c₈ + 331//80*c₉
 -1//720*c₁^3*c₇ + 1//240*c₁*c₂*c₇ + 1//720*c₁^2*c₈ - 1//240*c₃*c₇ + 1//240*c₂*c₈ + 379//720*c₁*c₉ + 1087//144*c₁₀