Morphisms

Operators

IntersectionTheory.pullback — Function

pullback(f::AbsVarietyHom, x::ChRingElem)
pullback(f::AbsVarietyHom, F::AbsBundle)

Compute the pullback of a Chow ring element $x$ or a bundle $F$ by a morphism $f$.

IntersectionTheory.pushforward — Function

pushforward(f::AbsVarietyHom, x::ChRingElem)
pushforward(f::AbsVarietyHom, F::AbsBundle)

Compute the pushforward of a Chow ring element $x$ or a bundle $F$ by a morphism $f$. The pushforward of a bundle $F$ is understood as the alternating sum of all its direct images.

Base.:* — Method

*(f::AbsVarietyHom, g::AbsVarietyHom)

Construct the composition morphism $g\circ f: X\to Z$ for $f: X\to Y$ and $g:Y\to Z$.

AbstractAlgebra.Generic.dim — Method

dim(f::AbsVarietyHom)

Return the relative dimension.

IntersectionTheory.tangent_bundle — Method

tangent_bundle(f::AbsVarietyHom)

Return the relative tangent bundle.

IntersectionTheory.cotangent_bundle — Method

cotangent_bundle(f::AbsVarietyHom)

Return the relative cotangent bundle.

IntersectionTheory.todd — Method

todd(f::AbsVarietyHom)

Compute the Todd class of the relative tangent bundle.

Blowup

IntersectionTheory.blowup — Function

blowup(i::AbsVarietyHom)

Construct the blowup $\mathrm{Bl}_XY$ of an inclusion $i:X\hookrightarrow Y$.

Return the variety $\mathrm{Bl}$ and the exceptional divisor $E$.

Examples

julia> P = proj(2)
AbsVariety of dim 2

julia> Bl, E = blowup(point() → P)
(AbsVariety of dim 2, AbsVariety of dim 1)

julia> euler(Bl)
4

julia> integral(pushforward(E → Bl, E(1))^2)
-1