Varieties

Basic invariants

IntersectionTheory.degree — Function

degree(X::AbsVariety)

Compute the degree of $X$ with respect to its polarization.

IntersectionTheory.euler — Function

euler(X::AbsVariety)
euler(X::TnVariety)

Compute the Euler number of a variety $X$.

IntersectionTheory.hilbert_polynomial — Function

hilbert_polynomial(F::AbsBundle)
hilbert_polynomial(X::AbsVariety)

Compute the Hilbert polynomial of a bundle $F$ or the Hilbert polynomial of $X$ itself, with respect to the polarization $\mathcal O_X(1)$ on $X$.

IntersectionTheory.signature — Function

signature(X::AbsVariety)

Compute the signature of a variety $X$ of even dimension.

Examples

julia> degree(grassmannian(2, 4))
2

julia> euler(grassmannian(2, 4))
6

julia> hilbert_polynomial(proj(3))
1//6*t^3 + t^2 + 11//6*t + 1

julia> signature(complete_intersection(proj(3), 4)) # signature of a K3 surface
-16

Characteristic classes

By common abuse of notations, the Chern/Todd/... class of a variety will mean the Chern/Todd/... class of its tangent bundle.

IntersectionTheory.chern — Method

chern(X::AbsVariety)
chern(X::TnVariety)

Compute the total Chern class of the tangent bundle of $X$.

IntersectionTheory.chern — Method

chern(k::Int, X::AbsVariety)
chern(k::Int, X::TnVariety)

Compute the $k$-th Chern class of the tangent bundle of $X$.

IntersectionTheory.chern_number — Function

chern_number(X::AbsVariety, λ::Int...)
chern_number(X::AbsVariety, λ::Vector{Int})
chern_number(X::AbsVariety, λ::Partition)

Compute the Chern number $c_\lambda (X):=\int_X c_{\lambda_1}(X)\cdots c_{\lambda_k}(X)$, where $\lambda:=(\lambda_1,\dots,\lambda_k)$ is a partition of the dimension of $X$.

IntersectionTheory.chern_numbers — Function

chern_numbers(X::AbsVariety)

Compute all the Chern numbers of $X$ as a list of pairs $\lambda\Rightarrow c_\lambda(X)$.

IntersectionTheory.todd — Method

todd(X::AbsVariety)

Compute the Todd class of the tangent bundle of $X$.

IntersectionTheory.pontryagin — Method

pontryagin(X::AbsVariety)

Compute the total Pontryagin class of the tangent bundle of $X$.

IntersectionTheory.pontryagin — Method

pontryagin(k::Int, X::AbsVariety)

Compute the $k$-th Pontryagin class of the tangent bundle of $X$.

IntersectionTheory.a_hat_genus — Method

a_hat_genus(X::AbsVariety)

Compute the top $\hat A$ genus of a variety $X$ of even dimension.

IntersectionTheory.a_hat_genus — Method

a_hat_genus(k::Int, X::AbsVariety)

Compute the $k$-th $\hat A$ genus of a variety $X$.

IntersectionTheory.l_genus — Method

l_genus(X::AbsVariety)

Compute the top L genus of a variety $X$ of even dimension.

IntersectionTheory.l_genus — Method

l_genus(k::Int, X::AbsVariety)

Compute the $k$-th L genus of a variety $X$.

We also have the following functions for producing generic formulae.

IntersectionTheory.todd — Method

todd(n::Int)

Compute the (generic) $n$-th Todd genus.

IntersectionTheory.a_hat_genus — Method

a_hat_genus(n::Int)

Compute the (generic) $n$-th $\hat A$ genus.

IntersectionTheory.l_genus — Method

l_genus(n::Int)

Compute the (generic) $n$-th L genus.

Examples

julia> chern(proj(2))
3*h^2 + 3*h + 1

julia> todd(proj(2))
h^2 + 3//2*h + 1

julia> pontryagin(proj(2))
3*h^2 + 1

julia> chern_number(proj(3), [2,1])
24

julia> chern_numbers(proj(3))
3-element Vector{Pair{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq}}:
   3₁ => 4
 2₁1₁ => 24
   1₃ => 64

julia> l_genus(proj(2))
1

julia> a_hat_genus(proj(2))
-1//8

julia> todd(2)
1//12*c₁^2 + 1//12*c₂

julia> a_hat_genus(2)
7//5760*p₁^2 - 1//1440*p₂

julia> l_genus(2)
-1//45*p₁^2 + 7//45*p₂

Chow ring

AbstractAlgebra.Generic.basis — Function

basis(X::AbsVariety)

Return an additive basis of the Chow ring of $X$, grouped by increasing degree (i.e., increasing codimension).

basis(k::Int, X::AbsVariety)

Return an additive basis of the Chow ring of $X$ in codimension $k$.

Singular.betti — Function

betti(X::AbsVariety)

Return the Betti numbers of the Chow ring of $X$. Note that these are not necessarily equal to the usual Betti numbers, i.e., the dimensions of (co)homologies.

AbstractAlgebra.Generic.integral — Function

integral(x::ChRingElem)

Compute the integral of a Chow ring element.

If the variety $X$ has a (unique) point class X.point, the integral will be a number (an fmpq or a function field element). Otherwise the 0-dimensional part of $x$ is returned.

IntersectionTheory.dual_basis — Function

dual_basis(k::Int, X::AbsVariety)

Compute the dual basis of the additive basis in codimension $k$ given by basis(k, X) (the returned elements are therefore in codimension $\dim X-k$).

dual_basis(X::AbsVariety)

Compute the dual basis with respect to the additive basis given by basis(X), grouped by decreasing degree (i.e., decreasing codimension).

IntersectionTheory.intersection_matrix — Function

intersection_matrix(a::Vector)
intersection_matrix(a::Vector, b::Vector)
intersection_matrix(X::AbsVariety)

Compute the intersection matrix among entries of a vector $a$ of Chow ring elements, or between two vectors $a$ and $b$. For a variety $X$, this computes the intersection matrix of the additive basis given by basis(X).

Return a Nemo.fmpq_mat if possible, or a Matrix otherwise.

Methods for Grassmannians.

IntersectionTheory.schubert_class — Function

schubert_class(G::AbsVariety, λ::Int...)
schubert_class(G::AbsVariety, λ::Vector{Int})
schubert_class(G::AbsVariety, λ::Partition)

Return the Schubert class $\sigma_\lambda$ on a (relative) Grassmannian $G$.

IntersectionTheory.schubert_classes — Function

schubert_classes(m::Int, G::AbsVariety)

Return all the Schubert classes in codimension $m$ on a (relative) Grassmannian $G$.

Examples

julia> basis(proj(2))
3-element Vector{Vector{IntersectionTheory.ChRingElem}}:
 [1]
 [h]
 [h^2]

julia> G = grassmannian(2, 4)
AbsVariety of dim 4

julia> basis(G)
5-element Vector{Vector{IntersectionTheory.ChRingElem}}:
 [1]
 [c₁]
 [c₂, c₁^2]
 [c₁*c₂]
 [c₂^2]

julia> dual_basis(G)
5-element Vector{Vector{IntersectionTheory.ChRingElem}}:
 [c₂^2]
 [c₁*c₂]
 [-c₁^2 + 2*c₂, c₁^2 - c₂]
 [c₁]
 [1]

julia> betti(G)
5-element Vector{Int64}:
 1
 1
 2
 1
 1

julia> basis(2, G)
2-element Vector{IntersectionTheory.ChRingElem}:
 c₂
 c₁^2

julia> intersection_matrix(basis(2, G))
[1   1]
[1   2]

julia> intersection_matrix(basis(2, G), dual_basis(2, G))
[1   0]
[0   1]

julia> intersection_matrix(G)
[0   0   0   0   0   1]
[0   0   0   0   1   0]
[0   0   1   1   0   0]
[0   0   1   2   0   0]
[0   1   0   0   0   0]
[1   0   0   0   0   0]

julia> schubert_class(G, [1,1])
c₂

julia> schubert_classes(2, G)
2-element Vector{IntersectionTheory.ChRingElem}:
 c₁^2 - c₂
 c₂

Notice that intersection_matrix(X) contains a lot of redundant information: intersection numbers that are not in complementary codimensions are always 0. So usually it is better to separate the components in different codimensions.

Construct new varieties

See also Constructors.

Base.:* — Method

*(X::AbsVariety, Y::AbsVariety)

Construct the product variety $X\times Y$. If both $X$ and $Y$ have a polarization, $X\times Y$ will be endowed with the polarization of the Segre embedding.

IntersectionTheory.complete_intersection — Function

complete_intersection(X::AbsVariety, degs::Int...)
complete_intersection(X::AbsVariety, degs::Vector{Int})

Construct the complete intersection in $X$ of general hypersurfaces with degrees $d_1,\dots,d_k$.

IntersectionTheory.section_zero_locus — Function

section_zero_locus(F::AbsBundle)

Construct the zero locus of a general section of a bundle $F$.

Use the argument class=true to only compute the class of the zero locus (same as ctop(F)).

IntersectionTheory.degeneracy_locus — Function

degeneracy_locus(k::Int, F::AbsBundle, G::AbsBundle)

Construct the $k$-degeneracy locus for a general bundle map from $F$ to $G$.

Use the argument class=true to only compute the class of the degeneracy locus.

Examples

julia> proj(1) * proj(1)
AbsVariety of dim 2

julia> complete_intersection(proj(3), 4)
AbsVariety of dim 2

julia> section_zero_locus(OO(proj(3), 4))
AbsVariety of dim 2

julia> X, (A, B) = variety(2, [3=>"a", 4=>"b"]);

julia> D = degeneracy_locus(2, A, B)
AbsVariety of dim 0

julia> pushforward(D → X, D(1))
a₁^2 - a₁*b₁ - a₂ + b₂

julia> degeneracy_locus(2, A, B, class=true)
a₁^2 - a₁*b₁ - a₂ + b₂