Chern number computations

We gather some Chern numbers related computations here.

Chern numbers of the Hilbert scheme of a surface

Let $S$ be a smooth projective surface. Ellingsrud–Göttsche–Lehn showed that the cobordism class of the Hilbert scheme $S^{[n]}$ of $n$ points on $S$ depends only on the cobordism class of $S$. In other words, the Chern numbers of $S^{[n]}$ only depends on the Chern numbers of $S$, $c_1(S)^2$ and $c_2(S)$.

On the other hand, the cobordism classes of surfaces are linearly generated by $[\mathbf P^2]$ and $[\mathbf P^1\times \mathbf P^1]$. For these two surfaces, their Hilbert schemes admit natural torus actions, so their Chern numbers can be readily computed using Bott's formula.

We provide the constructors for $(\mathbf P^2)^{[n]}$ and $(\mathbf P^1\times \mathbf P^1)^{[n]}$ as TnVariety, which will allow the computation of their Chern numbers using Bott's formula. Then we implement a function that computes the Chern numbers of $S^{[n]}$ for an arbitrary cobordism class $[S]$ specified by its two Chern numbers $c_1^2$ and $c_2$, using the strategy given in loc. cit.

IntersectionTheory.hilb_P2 — Function

hilb_P2(n::Int)

Construct the Hilbert scheme of $n$ points on $\mathbf P^2$ as a TnVariety.

IntersectionTheory.hilb_P1xP1 — Function

hilb_P1xP1(n::Int)

Construct the Hilbert scheme of $n$ points on $\mathbf P^1\times\mathbf P^1$ as a TnVariety.

IntersectionTheory.hilb_surface — Function

hilb_surface(n::Int, c1_2::Int, c2::Int)
hilb_surface(n::Int, c1_2::Int, c2::Int, P::Vector{Partition})

Compute the Chern numbers of the Hilbert scheme of $n$ points on a surface with given Chern numbers $c_1^2$ and $c_2$.

IntersectionTheory.hilb_K3 — Function

hilb_K3(n::Int)

Compute the (non-trivial) Chern numbers of a hyperkähler variety of $\mathrm{K3}^{[n]}$-type.

Examples

julia> X = hilb_P2(2)
TnVariety of dim 4

julia> chern_numbers(X)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 5 entries:
  2₂   => 90
  2₁1₂ => 162
  3₁1₁ => 54
  1₄   => 243
  4₁   => 9

julia> hilb_surface(2, 0, 24)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 5 entries:
  2₂   => 828
  2₁1₂ => 0
  3₁1₁ => 0
  1₄   => 0
  4₁   => 324

julia> hilb_K3(2)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 2 entries:
  2₂ => 828
  4₁ => 324

We record the Chern numbers of $\mathrm{K3}^{[n]}$ with $n\in\{5,6,7\}$, since these are not found in the literature.

julia> hilb_K3(5)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 7 entries:
  10₁  => 176256
  4₁2₃ => 52697088
  4₂2₁ => 21921408
  6₁4₁ => 5075424
  6₁2₂ => 12168576
  2₅   => 126867456
  8₁2₁ => 1774080

julia> hilb_K3(6)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 11 entries:
  6₁4₁2₁ => 392090040
  8₁4₁   => 59314272
  8₁2₂   => 139942280
  12₁    => 1073720
  6₁2₃   => 927397840
  10₁2₁  => 14450680
  4₃     => 697106648
  2₆     => 9277276480
  4₁2₄   => 3910848640
  4₂2₂   => 1650311720
  6₂     => 93495320

julia> hilb_K3(7)
Dict{AbstractAlgebra.Generic.Partition{Int64}, Nemo.fmpq} with 15 entries:
  8₁6₁   => 1296158976
  10₁4₁  => 569044224
  10₁2₂  => 1324608768
  6₁2₄   => 79324710912
  6₁4₂   => 14528215296
  2₇     => 765374164992
  4₃2₁   => 59674012416
  14₁    => 5930496
  8₁4₁2₁ => 5296568832
  6₂2₁   => 8273055744
  12₁2₁  => 102477312
  8₁2₃   => 12357114624
  6₁4₁2₂ => 33935583744
  4₂2₃   => 139582386432
  4₁2₅   => 326732507136

A polynomial defined by Libgober–Wood

Libgober–Wood defined a universal polynomial with coefficients in Chern numbers of an $n$-dimensional compact Kähler manifold, which can be used to obtain restrictions on certain topological invariants. They are computed up to dimension 9 in this survey.

We have a function to compute this polynomial. The dimension-10 one can be computed in less than 1 second, and the dimension-16 one in about 1 minute.

Of course one can also compute the value of this polynomial for a specific variety.

IntersectionTheory.libgober_wood_polynomial — Function

libgober_wood_polynomial(n::Int)
libgober_wood_polynomial(X::AbsVariety)

Compute the polynomial defined by Libgober–Wood in dimension $n$.

Examples

julia> libgober_wood_polynomial(3)
-1//24*c₁*c₂*z^3 + (1//12*c₁*c₂ + 1//2*c₃)*z^2 - 3//2*c₃*z + c₃

julia> libgober_wood_polynomial(proj(6))
z^6 - 7*z^5 + 21*z^4 - 35*z^3 + 35*z^2 - 21*z + 7

The degree-2 term is shown to depend on only two Chern numbers.

julia> [Nemo.coeff(libgober_wood_polynomial(n), 2) for n in 1:10]
10-element Vector{IntersectionTheory.ChRingElem}:
 0
 1//12*c₁^2 + 1//12*c₂
 1//12*c₁*c₂ + 1//2*c₃
 1//12*c₁*c₃ + 7//6*c₄
 1//12*c₁*c₄ + 25//12*c₅
 1//12*c₁*c₅ + 13//4*c₆
 1//12*c₁*c₆ + 14//3*c₇
 1//12*c₁*c₇ + 19//3*c₈
 1//12*c₁*c₈ + 33//4*c₉
 1//12*c₁*c₉ + 125//12*c₁₀

We observe that there should be similar results for the degree-4 term.

julia> [Nemo.coeff(libgober_wood_polynomial(n), 4) for n in 1:10]
10-element Vector{IntersectionTheory.ChRingElem}:
 0
 0
 0
 -1//720*c₁^4 + 1//180*c₁^2*c₂ + 1//240*c₂^2 + 1//720*c₁*c₃ - 1//720*c₄
 -1//720*c₁^3*c₂ + 1//240*c₁*c₂^2 + 1//720*c₁^2*c₃ + 29//720*c₁*c₄ + 1//24*c₅
 -1//720*c₁^3*c₃ + 1//240*c₁*c₂*c₃ + 1//720*c₁^2*c₄ - 1//240*c₃^2 + 1//240*c₂*c₄ + 23//240*c₁*c₅ + 31//120*c₆
 -1//720*c₁^3*c₄ + 1//240*c₁*c₂*c₄ + 1//720*c₁^2*c₅ - 1//240*c₃*c₄ + 1//240*c₂*c₅ + 31//180*c₁*c₆ + 301//360*c₇
 -1//720*c₁^3*c₅ + 1//240*c₁*c₂*c₅ + 1//720*c₁^2*c₆ - 1//240*c₃*c₅ + 1//240*c₂*c₆ + 97//360*c₁*c₇ + 81//40*c₈
 -1//720*c₁^3*c₆ + 1//240*c₁*c₂*c₆ + 1//720*c₁^2*c₇ - 1//240*c₃*c₆ + 1//240*c₂*c₇ + 31//80*c₁*c₈ + 331//80*c₉
 -1//720*c₁^3*c₇ + 1//240*c₁*c₂*c₇ + 1//720*c₁^2*c₈ - 1//240*c₃*c₇ + 1//240*c₂*c₈ + 379//720*c₁*c₉ + 1087//144*c₁₀